+++ core-ng-search/src/test/java/core/framework/search/impl/ElasticSearchMetricsTest.java
@@ -4,6 +4,8 @@ import core.framework.internal.stat.Stats;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import java.util.HashMap;
+
 import static org.assertj.core.api.Assertions.assertThat;
 
 class ElasticSearchMetricsTest {
@@ -21,7 +23,7 @@ class ElasticSearchMetricsTest {
     void collect() {
         instrumentation.activeRequests.increase();
 
-        var stats = new Stats();
+        var stats = new Stats(new HashMap<>());
         metrics.collect(stats);
 
         assertThat(stats.stats).containsEntry("es_active_requests", 1.0);