+++ core-ng/src/test/java/core/framework/internal/resource/PoolMetricsTest.java
@@ -4,6 +4,8 @@ import core.framework.internal.stat.Stats;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import java.util.HashMap;
+
 import static org.assertj.core.api.Assertions.assertThat;
 
 /**
@@ -29,7 +31,7 @@ class PoolMetricsTest {
         pool.borrowItem();
         pool.returnItem(pool.borrowItem());
 
-        var stats = new Stats();
+        var stats = new Stats(new HashMap<>());
         metrics.collect(stats);
 
         assertThat(stats.stats)